Avoid selenium-webdriver 4.5
In selenium-webdriver 4.5, the :debugger_address option's default was
changed to true. This leads to an error stating that moz:debuggerAddress
is an unknown option. Since it's hard to set the :debugger_address
option back to false when using the standard Rails system testing
machinery, the quick fix is to exclude version 4.5 of the
selenium-webdriver dependency.